pyproject.toml
[project]
name = "llm-daydreamer"
version = "0.1.0"
description = "Prototype implementation of a daydreaming loop around an LLM"
authors = [{name = "AutoGenerated", email = "n/a"}]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.21.0",
    "numpy>=1.26",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ['.']
